agrego builder del service

+++ TpIntegrador/src/main/java/ungs/utils/TimeUtils.java
@@ -4,7 +4,7 @@ public class TimeUtils {
 
     public static Long getMinutes(Long value) {
         if (value != null) {
-            return (long) value*1000*60;
+            return value * 1000 * 60;
         }
         throw new IllegalArgumentException("Value Not Valid");
     }
